tacy_app/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==4.1.2
django-cors-headers==3.13.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
drf-yasg==1.21.4
gunicorn==20.1.0
httpie==3.2.1
idna==3.4
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
packaging==22.0
Pillow==9.2.0
psycopg2-binary==2.9.5
Pygments==2.13.0
PySocks==1.7.1
pytz==2022.4
requests==2.28.1
requests-toolbelt==0.10.0
rich==12.6.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
typing-extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12